Circular dependencies fix
