+++ lib/index.js
@@ -8,7 +8,7 @@
 import {html, svg} from 'property-information'
 import {htmlVoidElements} from 'html-void-elements'
 import {omission} from './omission/index.js'
-import {one} from './one.js'
+import {one} from './tree.js'
 
 /**
  * @param {Node|Array.<Node>} node
