+++ lib/tree.js
@@ -4,6 +4,7 @@
  * @typedef {import('./types.js').Context} Context
  * @typedef {import('./types.js').Properties} Properties
  * @typedef {import('./types.js').PropertyValue} PropertyValue
+ * @typedef {import('./types.js').Parent} Parent
  */
 
 import {svg, find} from 'property-information'
@@ -11,8 +12,61 @@ import {stringify as spaces} from 'space-separated-tokens'
 import {stringify as commas} from 'comma-separated-tokens'
 import {stringifyEntities} from 'stringify-entities'
 import {ccount} from 'ccount'
-import {all} from './all.js'
 import {constants} from './constants.js'
+import {comment} from './comment'
+import {doctype} from './doctype'
+import {raw} from './raw'
+import {text} from './text'
+
+/**
+ * @type {Object.<string, Handle>}
+ */
+var handlers = {
+  comment,
+  doctype,
+  element,
+  // @ts-ignore `raw` is nonstandard
+  raw,
+  // @ts-ignore `root` is a parent.
+  root: all,
+  text
+}
+
+var own = {}.hasOwnProperty
+
+/**
+ * @type {Handle}
+ */
+export function one(ctx, node, index, parent) {
+  if (!node || !node.type) {
+    throw new Error('Expected node, not `' + node + '`')
+  }
+
+  if (!own.call(handlers, node.type)) {
+    throw new Error('Cannot compile unknown node `' + node.type + '`')
+  }
+
+  return handlers[node.type](ctx, node, index, parent)
+}
+
+/**
+ * Serialize all children of `parent`.
+ *
+ * @type {Handle}
+ * @param {Parent} parent
+ */
+export function all(ctx, parent) {
+  /** @type {Array.<string>} */
+  var results = []
+  var children = (parent && parent.children) || []
+  var index = -1
+
+  while (++index < children.length) {
+    results[index] = one(ctx, children[index], index, parent)
+  }
+
+  return results.join('')
+}
 
 /**
  * @type {Handle}
